+++ lib/logidze.rb
@@ -7,6 +7,7 @@ module Logidze
   require 'logidze/history'
   require 'logidze/model'
   require 'logidze/versioned_association'
+  require 'logidze/ignore_log_data'
   require 'logidze/has_logidze'
   require 'logidze/meta'