+++ spec/dummy/app/models/user.rb
@@ -3,6 +3,8 @@ class User < ActiveRecord::Base
 
   delegate :responsible_id, :meta, to: :log_data
 
+  has_many :not_logged_posts
+
   def whodunnit
     User.find(responsible_id) if responsible_id.present?
   end